wellbee/requirements.txt
asgiref==3.8.1
boto3==1.35.49
botocore==1.35.49
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==4.2.14
django-cors-headers==4.4.0
django-environ==0.12.0
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
gunicorn==23.0.0
idna==3.7
jmespath==1.0.1
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
Pillow==10.0.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.3
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
twilio==9.4.5
firebase-admin==6.5.0
phonenumbers==8.13.47
typing_extensions==4.12.2
urllib3==1.26.20
